refactor dpm campaign discovery route
